+++ src/lein_monolith/plugin.clj
@@ -199,7 +199,6 @@
       (update subproject :middleware (fnil conj []) mw-sym))))
 
 
-
 ;; ## Merged Profiles
 
 (def ^:private path-keys